fix(core): address code review issues in CLI and agent meta-tools
- Gate flush_pending_tool_calls behind both ai-providers and gui
  features to prevent compilation failure without Tauri
- Add topological sort for plan step execution to honor depends_on
  ordering with cycle detection (Kahn's algorithm)
- Report rollback failures instead of silently discarding undo errors
- Add cycle detection to plan validate command
- Honor --sequence argument in render start command
- Remove duplicate trackCount key in state dump JSON output
- Validate unsupported format parameter in asset list command
- Use f32 directly for speed CLI argument to avoid f64-to-f32 cast
- Block execute_plan self-invocation to prevent unbounded recursion
- Add fallback in BackendToolExecutor when meta-tools are enabled but
  not yet registered

+++ crates/openreelio-cli/src/commands/timeline.rs
@@ -179,7 +179,7 @@ pub enum TimelineAction {
 
         /// Speed multiplier (e.g. 2.0 for 2x)
         #[arg(long)]
-        speed: f64,
+        speed: f32,
 
         /// Reverse playback
         #[arg(long, default_value = "false")]
@@ -445,7 +445,7 @@ pub async fn execute(action: TimelineAction) -> anyhow::Result<()> {
         TimelineAction::Speed { path, clip, track, speed, reverse, sequence } => {
             let mut project = super::load_project(&path)?;
             let seq_id = resolve_sequence_id(&project, sequence)?;
-            let cmd = SetClipSpeedCommand::new(&seq_id, &track, &clip, speed as f32, reverse);
+            let cmd = SetClipSpeedCommand::new(&seq_id, &track, &clip, speed, reverse);
             let result = project
                 .executor
                 .execute(Box::new(cmd), &mut project.state)
